Fixed race condition when multiple SETUP messages queued.
When USB_Device_ProcessControlRequest() is called, it first does the HID-specific EVENT_USB_Device_ControlRequest() processing.  If that doesn't handle the request, then since Endpoint_isSETUPReceived() is still true, then to do the generic processing.  If _that_ doesn't handle the request, then it calls Endpoint_ClearSETUP() and stalls the transaction.

In some cases, a new SETUP message is arriving between the Endpoint_ClearSETUP() in EVENT_USB_Device_ControlRequest() and the check for Endpoint_isSETUPReceived() right after EVENT_USB_Device_ControlRequest() returns.

In this case, the switch statement in USB_Device_ProcessControlRequest() tries to process the HID-specific message with the generic processing.  It doesn't handle the HID-specific message, so it falls through to the STALL code.

My suggestion is to only check for Endpoint_isSETUPReceived() once at the top of USB_Device_ProcessControlRequest().  If it is set, then try first the specific handler, then if that doesn't satisfy the request, then try the generic handler.  If _that_ doesn't handle the request, then generate the STALL.  I did this by using a request_handled variable, and modifying EVENT_USB_Device_ControlRequest() to return a value that says whether it handled the request or not.